Invalid Redirect Problem
Using _PS_BASE_URL_ . __PS_BASE_URI__ will not return "https://" at the start of the URL if your store is not set to use HTTPS for every page (most people don't have HTTPS enabled for every page as it is unnecessary/a waste of resources). The problem with this is that when you use the configuration page to get the API token, GoCoin will say the URL to redirect to is invalid as the URL must start with "https".