Error caused by invalid comment
Some HTML has been commented out using <!-- and -->, which is fine for stopping the elements from actually appearing on the page, however the condition in the if statement is still being checked, which is what's triggering an error, as the variable does not exist. Instead, the whole block should've been commented out using {* and *}, which is how to comment out things in the smarty templating system.